+++ lib/authentication/view/email_code_verification_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_news_app_mobile_client_full_source_code/app/bloc/app_bloc.dart';
 import 'package:flutter_news_app_mobile_client_full_source_code/app/config/config.dart';
 import 'package:flutter_news_app_mobile_client_full_source_code/authentication/bloc/authentication_bloc.dart';
 import 'package:flutter_news_app_mobile_client_full_source_code/l10n/l10n.dart';
@@ -74,9 +73,9 @@ class EmailCodeVerificationPage extends StatelessWidget {
                         textAlign: TextAlign.center,
                       ),
                       // Display demo code if in demo environment
-                      BlocSelector<AppBloc, AppState, AppEnvironment?>(
-                        selector: (state) => state.environment,
-                        builder: (context, environment) {
+                      Builder(
+                        builder: (context) {
+                          final environment = context.read<AppEnvironment>();
                           if (environment == AppEnvironment.demo) {
                             return Column(
                               children: [