+++ lib/shared/extensions/extensions.dart
@@ -1,1 +1,4 @@
+export 'app_status_x.dart';
+export 'app_user_x.dart';
 export 'content_type_extensions.dart';
+export 'user_feed_decorator_status_x.dart';